+++ brkt_cli/brkt_jwt/test_jwt.py
@@ -25,8 +25,8 @@ import tempfile
 import time
 import unittest
 
-import brkt_cli.jwt
-import brkt_cli.jwt.jwk
+import brkt_cli.brkt_jwt
+import brkt_cli.brkt_jwt.jwk
 
 
 _signing_key = SigningKey.generate(curve=NIST384p)
@@ -37,17 +37,17 @@ class TestTimestamp(unittest.TestCase):
     def test_datetime_to_timestamp(self):
         now = time.time()
         dt = datetime.fromtimestamp(now, tz=iso8601.UTC)
-        self.assertEqual(now, brkt_cli.jwt._datetime_to_timestamp(dt))
+        self.assertEqual(now, brkt_cli.brkt_jwt._datetime_to_timestamp(dt))
 
     def test_parse_timestamp(self):
         ts = int(time.time())
         dt = datetime.fromtimestamp(ts, tz=iso8601.UTC)
 
-        self.assertEqual(dt, brkt_cli.jwt.parse_timestamp(str(ts)))
-        self.assertEqual(dt, brkt_cli.jwt.parse_timestamp(dt.isoformat()))
+        self.assertEqual(dt, brkt_cli.brkt_jwt.parse_timestamp(str(ts)))
+        self.assertEqual(dt, brkt_cli.brkt_jwt.parse_timestamp(dt.isoformat()))
 
         with self.assertRaises(ValidationError):
-            brkt_cli.jwt.parse_timestamp('abc')
+            brkt_cli.brkt_jwt.parse_timestamp('abc')
 
 
 class TestSigningKey(unittest.TestCase):
@@ -60,7 +60,7 @@ class TestSigningKey(unittest.TestCase):
         key_file.write(pem)
         key_file.flush()
 
-        signing_key = brkt_cli.jwt.read_signing_key(key_file.name)
+        signing_key = brkt_cli.brkt_jwt.read_signing_key(key_file.name)
         self.assertEqual(pem, signing_key.to_pem())
         key_file.close()
 
@@ -74,7 +74,7 @@ class TestSigningKey(unittest.TestCase):
         key_file.flush()
 
         with self.assertRaises(ValidationError):
-            brkt_cli.jwt.read_signing_key(key_file.name)
+            brkt_cli.brkt_jwt.read_signing_key(key_file.name)
         key_file.close()
 
     def test_read_signing_key_io_error(self):
@@ -82,11 +82,11 @@ class TestSigningKey(unittest.TestCase):
         """
         # Read from a directory.
         with self.assertRaises(ValidationError):
-            brkt_cli.jwt.read_signing_key('.')
+            brkt_cli.brkt_jwt.read_signing_key('.')
 
         # Read from a file that doesn't exist.
         with self.assertRaises(ValidationError):
-            brkt_cli.jwt.read_signing_key('nothing_here.pem')
+            brkt_cli.brkt_jwt.read_signing_key('nothing_here.pem')
 
         # Read from a malformed file.
         key_file = tempfile.NamedTemporaryFile()
@@ -94,7 +94,7 @@ class TestSigningKey(unittest.TestCase):
         key_file.flush()
 
         with self.assertRaises(ValidationError):
-            brkt_cli.jwt.read_signing_key(key_file.name)
+            brkt_cli.brkt_jwt.read_signing_key(key_file.name)
         key_file.close()
 
 
@@ -111,7 +111,7 @@ class TestGenerateJWT(unittest.TestCase):
         cnc = 10
         customer = str(uuid.uuid4())
 
-        jwt = brkt_cli.jwt.make_jwt(
+        jwt = brkt_cli.brkt_jwt.make_jwt(
             _signing_key,
             nbf=nbf,
             exp=exp,
@@ -142,13 +142,13 @@ class TestGenerateJWT(unittest.TestCase):
         self.assertEqual(1, payload['one'])
         self.assertEqual(2, payload['two'])
 
-        iat = brkt_cli.jwt._timestamp_to_datetime(payload['iat'])
+        iat = brkt_cli.brkt_jwt._timestamp_to_datetime(payload['iat'])
         self.assertTrue(now <= iat <= after)
 
-        nbf_result = brkt_cli.jwt._timestamp_to_datetime(payload['nbf'])
+        nbf_result = brkt_cli.brkt_jwt._timestamp_to_datetime(payload['nbf'])
         self.assertEqual(nbf, nbf_result)
 
-        exp_result = brkt_cli.jwt._timestamp_to_datetime(payload['exp'])
+        exp_result = brkt_cli.brkt_jwt._timestamp_to_datetime(payload['exp'])
         self.assertEqual(exp, exp_result)
 
         # Check signature.
@@ -163,7 +163,7 @@ class TestGenerateJWT(unittest.TestCase):
         """ Test that claims specified by name are embedded into the JWT. """
         # Generate the JWT.
         claims = {'foo': 'bar', 'count': 5}
-        jwt = brkt_cli.jwt.make_jwt(_signing_key, claims=claims)
+        jwt = brkt_cli.brkt_jwt.make_jwt(_signing_key, claims=claims)
         _, payload_b64, _ = jwt.split('.')
         payload_json = brkt_cli.util.urlsafe_b64decode(payload_b64)
         payload = json.loads(payload_json)
@@ -176,5 +176,5 @@ class TestJWK(unittest.TestCase):
 
     def test_long_to_byte_array(self):
         l = long('deadbeef', 16)
-        ba = brkt_cli.jwt.jwk._long_to_byte_array(l)
+        ba = brkt_cli.brkt_jwt.jwk._long_to_byte_array(l)
         self.assertEqual(bytearray.fromhex('deadbeef'), ba)
